+++ Models/AppSettings.cs
@@ -11,7 +11,9 @@ public class AppSettings
     public bool MinimizeToTray { get; set; } = true;
     public bool CloseMinimizesToTray { get; set; } = true;
     public bool AutoCheckForUpdates { get; set; } = true;
+    public bool IsVirtualOutputActive { get; set; } = true;
     public bool IsHidHideActive { get; set; } = false;
     public bool IsAppListInverted { get; set; } = false;
     public List<string> BlockedInstanceIds { get; set; } = new();
+    public List<string> WhitelistedApplications { get; set; } = new();
 }